# docs/swagger.yaml
basePath: /
definitions:
  handlers.ConvertLeadRequest:
    properties:
      amount:
        example: "50000"
        type: string
      currency:
        example: USD
        type: string
    type: object
  models.Deals:
    properties:
      amount:
        type: string
      created_at:
        type: string
      currency:
        type: string
      id:
        type: integer
      lead_id:
        type: integer
      status:
        type: string
    type: object
...
  models.Leads:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      owner_id:
        type: integer
      status:
        type: string
      title:
        type: string
    type: object
...
host: localhost:4000
info:
  contact:
    email: support@company.com
    name: Тур Компани
    url: http://contact.company.com
  description: Aпишка для управления туристической компанией (документация Swagger).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://your-terms.com/terms/
  title: Tour Company API
  version: "1.0"
paths:
...
  /leads:
    post:
      consumes:
      - application/json
      description: Создает нового клиента (лида)
      parameters:
      - description: Данные нового лида
        in: body
        name: lead
        required: true
        schema:
          $ref: '#/definitions/models.Leads'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Leads'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать лид
      tags:
      - Leads
  /leads/:
    get:
      description: Returns a list of all leads
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Leads'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all leads
      tags:
      - Leads
  /leads/{id}:
    delete:
      description: Удаляет клиента по ID
      parameters:
      - description: ID Лида
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить лида
      tags:
      - Leads
    get:
      description: Возвращает данные одного лида
      parameters:
      - description: ID Лида
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Leads'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить лид по ID
      tags:
      - Leads
    put:
      consumes:
      - application/json
      description: Обновляет данные лида по ID
      parameters:
      - description: ID Лида
        in: path
        name: id
        required: true
        type: integer
      - description: Обновленные данные
        in: body
        name: lead
        required: true
        schema:
          $ref: '#/definitions/models.Leads'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Leads'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить лид
      tags:
      - Leads
  /leads/{id}/convert:
    put:
      consumes:
      - application/json
      description: Создает сделку на основе существующего лида
      parameters:
      - description: ID лида
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для сделки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ConvertLeadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Deals'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Конвертировать лид в сделку
      tags:
      - Leads
...
  /reports/deals/filter:
    get:
      description: Фильтрует сделки по статусу и дате.
      parameters:
      - description: Статус
        in: query
        name: status
        type: string
      - description: Дата с (yyyy-mm-dd)
        in: query
        name: from
        type: string
      - description: Дата по (yyyy-mm-dd)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Deals'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Фильтрация сделок
      tags:
      - Reports
  /reports/leads/filter:
    get:
      description: Фильтрует лиды по статусу и owner_id.
      parameters:
      - description: Статус
        in: query
        name: status
        type: string
      - description: ID владельца
        in: query
        name: owner_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Leads'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Фильтрация лидов
      tags:
      - Reports
  /reports/summary:
    get:
      description: Выводит общее количество лидов и сделок.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Сводный отчет
      tags:
      - Reports
...
swagger: "2.0"